# Optimize Element.OnPropertyChanged