- Don't foreach over the empty collection ChildrenNotDrawnByThisElement. The bulk of Element subclasses don't provide a value. Instead just return null.

This is a pretty minor, low-level optimization. But it shows to save 10% off setting simple properties.

## Benchmark Results

### Before

|      Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|----------:|----------:|------:|------:|------:|----------:|
| SetProperty | 1.736 ms | 0.0068 ms | 0.0057 ms |     - |     - |     - |     665 B |

### After

|      Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|----------:|----------:|------:|------:|------:|----------:|
| SetProperty | 1.546 ms | 0.0151 ms | 0.0133 ms |     - |     - |     - |     665 B |